Byiers et al., (2014)

Data from an ABAB design that evaluated the effectiveness of functional assessment and functional communication training on the Rett syndrome (RTT) of three individuals with RTT novel communicative behaviors. The dependent variable, communicative behavior, was measured using frequency counting for a five minute observation session. A single outcome (percent of switch activated responses) was calculated with appropriate responses per minute divided by inappropriate responses per minute, multiplied by 100. The variables are as follows:

StudyID. Study identifier.

StudyID_CaseID. Participant identifier.

Session_number. Within-series session-number.

Condition. Indicates condition, (A for baseline, B for treatment).

Outcome. Value for the percent of switch activated responses.

Session_length. Length of the session.

Procedure. Scale of the outcome variable.

Goal_level. Indicates the goal level for each series.

phase_pair_calculated. Indicates the unique phase number of each sequential occurrence of a phase for each participant of ABAB design.

Format

A data frame with 54 rows and 9 variables

Source

Byiers, B. J., Dimian, A., & Symons, F. J. (2014). Functional communication training in Rett syndrome: A preliminary study. American Journal on Intellectual and Developmental Disabilities, 119(4), 340-350. doi:10.1352/1944-7558-119.4.340