Percentage of all non-overlapping data (PAND)

Calculates the percentage of all non-overlapping data index (Parker, Hagan-Burke, & Vannest, 2007; Parker, Vannest, & Davis, 2011).

PAND(
  A_data,
  B_data,
  condition,
  outcome,
  baseline_phase = NULL,
  intervention_phase = NULL,
  improvement = "increase"
)

Arguments

A_data

vector of numeric data for A phase. Missing values are dropped.

B_data

vector of numeric data for B phase. Missing values are dropped.

condition

vector identifying the treatment condition for each observation in the series.

outcome

vector of outcome data for the entire series.

baseline_phase

character string specifying which value of condition corresponds to the baseline phase. Defaults to first observed value of condition.

intervention_phase

character string specifying which value of condition corresponds to the intervention phase. Defaults to second unique value of condition.

improvement

character string indicating direction of improvement. Default is "increase".

Value

Numeric value

Details

For an outcome where increase is desirable, PAND is calculated as the proportion of observations remaining after removing the fewest possible number of observations from either phase so that the highest remaining point from the baseline phase is less than the lowest remaining point from the treatment phase. For an outcome where decrease is desirable, PAND is calculated as the proportion of observations remaining after removing the fewest possible number of observations from either phase so that the lowest remaining point from the baseline phase is greater than the highest remaining point from the treatment phase. The range of PAND depends on the number of observations in each phase.

References

Parker, R. I., Hagan-Burke, S., & Vannest, K. J. (2007). Percentage of all non-overlapping data (PAND): An alternative to PND. The Journal of Special Education, 40(4), 194--204. doi:doi:10.1177/00224669070400040101

Parker, R. I., Vannest, K. J., & Davis, J. L. (2011). Effect size in single-case research: A review of nine nonoverlap techniques. Behavior Modification, 35(4), 303--22. doi:doi:10.1177/0145445511399147

Examples

A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
PAND(A_data = A, B_data = B)
#>     ES       Est
#> 1 PAND 0.8461538