Percentage of non-overlapping data

Calculates the percentage of non-overlapping data index (Scruggs, Mastropieri, & Castro, 1987).

PND(
  A_data,
  B_data,
  condition,
  outcome,
  baseline_phase = NULL,
  intervention_phase = NULL,
  improvement = "increase"
)

Arguments

A_data
vector of numeric data for A phase. Missing values are dropped.

B_data
vector of numeric data for B phase. Missing values are dropped.

condition
vector identifying the treatment condition for each observation in the series.

outcome
vector of outcome data for the entire series.

baseline_phase
character string specifying which value of condition corresponds to the baseline phase. Defaults to first observed value of condition.

intervention_phase
character string specifying which value of condition corresponds to the intervention phase. Defaults to second unique value of condition.

improvement
character string indicating direction of improvement. Default is "increase".

Value

Numeric value

Details

For an outcome where increase is desirable, PND is calculated as the proportion of observations in the B phase that exceed the highest observation from the A phase. For an outcome where decrease is desirable, PND is the proportion of observations in the B phase that are less than the lowest observation from the A phase. The range of PND is [0,1].

References

Scruggs, T. E., Mastropieri, M. A., & Casto, G. (1987). The quantitative synthesis of single-subject research: Methodology and validation. Remedial and Special Education, 8(2), 24--43. doi:doi: 10.1177/074193258700800206

Examples

A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
PND(A_data = A, B_data = B)
#>    ES       Est
#> 1 PND 0.7142857