Percent of Goal Obtained

Calculates the percent of goal obtained effect size index

PoGO(
  A_data,
  B_data,
  condition,
  outcome,
  goal,
  baseline_phase = NULL,
  intervention_phase = NULL,
  improvement = "increase",
  confidence = 0.95
)

Arguments

A_data

vector of numeric data for A phase. Missing values are dropped.

B_data

vector of numeric data for B phase. Missing values are dropped.

condition

vector identifying the treatment condition for each observation in the series.

outcome

vector of outcome data for the entire series.

goal

a numerical value indicating the goal level of behavior.

baseline_phase

character string specifying which value of condition corresponds to the baseline phase. Defaults to first observed value of condition.

intervention_phase

character string specifying which value of condition corresponds to the intervention phase. Defaults to second unique value of condition.

improvement

character string indicating direction of improvement. Default is "increase".

confidence

confidence level for the reported interval estimate. Set to NULL to omit confidence interval calculations.

Value

A data frame containing the estimate, standard error, and confidence interval.

Details

The percent of goal obtained (PoGO) effect size parameter is defined as the ratio of the difference in the mean level of behavior during phase B versus during phase A to the difference between the goal level of behavior and the mean level of behavior during phase A, multiplied by 100.

The standard error of PoGO is calculated based on Dunlap and Silver's (1986) approximation for the standard error of a ratio. The confidence interval for LRR is based on a large-sample (z) approximation.

References

Dunlap, W. P., & Silver, N. C. (1986). Confidence intervals and standard error for ratios of normal variables. Behavior Research Methods, Instruments, & Computers, 18, 469-471. doi:doi:10.3758/BF03201412

Ferron, J., Goldstein, H., Olszewski, A., & Rohrer, L. (2020). Indexing effects in single-case experimental designs by estimating the percent of goal obtained. Evidence-Based Communication Assessment and Intervention, 14(1-2), 6-27. doi:doi:10.1080/17489539.2020.1732024

Patrona, E., Ferron, J., Olszewski, A., Kelley, E., & Goldstein, H. (2022). Effects of explicit vocabulary interventions for preschoolers: An exploratory application of the percent of goal obtained (PoGO) effect size metric. Journal of Speech, Language, and Hearing Research, forthcoming.

Examples

A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
PoGO(A_data = A, B_data = B, goal = 30)
#>     ES      Est       SE CI_lower CI_upper
#> 1 PoGO 66.88312 20.74513 26.22341 107.5428