Schmidt (2007)

Data from a study by Schmidt (2007). The variables are as follows:

Behavior_type. Outcome measure description (disruptive behavior or on task behavior).

Procedure. The observation recording procedure used to measure the outcome.

Metric. The metric in which the outcome measurement is expressed ("count" for natural counts; "percentage" for percentage of intervals)

Session_length. Length (in minutes) of the observation sessions

Interval_length. If an interval method was used, the length of the intervals (in seconds); NA otherwise.

Case_Pseudonym. Case Pseudonym provided by the authors.

Session_number. Within-series session-number.

Phase. Label for each unique phase (e.g., A1 is the first baseline phase, B2 is the second treatment phase).

Condition. Label indicating whether the outcome is in the baseline (A) or treatment (B) phase.

Outcome. Outcome measurement.

Phase_num. Indicator for each pair of baseline and treatment phases.

direction. Direction of therapeutic improvement for the outcome.

n_Intervals. If an interval method was used, the total number of intervals; NA otherwise.

Format

A data frame with 172 rows and 13 variables

Source

Schmidt, A. C. (2007). The effects of a group contingency on group and individual behavior in an urban first-grade classroom. Masters Thesis, University of Kansas, Department of Applied Behavioral Sciences. ProQuest Dissertations & Theses Global, thesis number 1443719. https://kuscholarworks.ku.edu/bitstream/handle/1808/32097/Schmidt_Anna_C_2007_5349292.pdf

Note

Data were reconstructed from figures in the source document. Consequently, outcome measurements in this dataset include additional error from the reconstruction process and may not exactly match the original source data.

In earlier versions of the package (0.5.0 and earlier), the Case_Pseudonym variable contained incorrect labels for each case, which did not match the labels in the source document. Specifically, Faith was incorrectly labeled as Albert; Lilly was incorrectly labeled as Faith; and Albert was incorrectly labeled as Lilly.