Schmidt and Stichter (2012)

Data from an ABAB design conducted by Schmidt and Stichter (2012). All data were collected via continuous recording. The variables are as follows:

Case. Participant identifier.

Behavior. Behavior type (Conversation, Initiations, or Responses).

Trt Treatment indicators.

Outcome. Outcome scores.

Session_num. Measurement occasion.

Session_length Length of the observation session.

Procedure The metric in which the outcome measurement is expressed, all "other".

Format

A data frame with 180 rows and 5 variables

Source

Schmidt, C., & Stichter, J. P. (2012). The use of peer-mediated interventions to promote the generalization of social competence for adolescents with high-functioning autism and Asperger's syndrome. Exceptionality, 20(2), 94-113. doi:10.1080/09362835.2012