Shogren et al. (2004)

Data from a systematic review by Shogren et al. (2004) on the effects of choice-making interventions. These data were compiled and re-analyzed in Pustejovsky (2015). The variables are as follows:

Study. An ID for each study in the systematic review.

Case. Case Pseudonym provided by the authors.

Measure. Type of behavior observed as the outcome measure

Phase. Phase indicator, baseline phase is "No Choice" and treatment phase is "Choice."

Percentage. For those outcomes measured as percentage, outcomes value. NA for count outcomes.

Observed. For those outcomes measured as count, outcome value. NA for percentage outcomes.

Possible. For counts out of a maximum, lists the maximum value.

Recording_procedure Recording procedure. CDR = "Continuous Duration Recording", EC = "Event Counting", "MTS = "Momentary Time Sampling", and PIR = "Partial Interval Recording."

Session_length. Length of the observation session in minutes.

interval_length. Length of the observation intervals for data observed using MTS or PIR.

outcome. Value for the outcome for all outcome types.

direction. Direction of therapeutic improvement for the outcome.

Format

A data frame with 634 rows and 15 variables

Source

Shogren, K. A., Faggella-Luby, M. N., Bae, S. J., & Wehmeyer, M. L. (2004). The effect of choice-making as an intervention for problem behavior. Journal of Positive Behavior Interventions, 6(4), 228–237.

References

Pustejovsky, J.E. (2015). Measurement-comparable effect sizes for single-case studies of free-operant behavior. Psychological Methods, 20(3), 342–359.