Thorne and Kamps (2008)

Data from an ABAB design conducted by Thorne and Kamps (2008). These data were used as an example in Swan and Pustejovsky (2017). Academic engagement was collected via continuous recording (marked as "other") and inappropriate verbalizations were collected via event counting (marked as "count"). The variables are as follows:

Measure. Outcome measure description (academic engagement or inappropriate verbalizations).

Case. Participant identifier.

Session_number. Measurement occasion.

Outcome. Outcome scores

Trt. Treatment indicators.

Session_length. Length of the observation session.

Measure. The metric in which the outcome measurement is expressed (count or other).

Format

A data frame with 776 rows and 5 variables

Source

Thorne, S., & Kamps, D. (2008). The effects of a group contingency intervention on academic engagement and problem behavior of at-risk students. Behavior Analysis in Practice, 1(2), 12-18.

References

Swan, D. M., & Pustejovsky, J. E. (2017). A gradual effects model for single-case designs. http://doi.org/10.17605/OSF.IO/GAXRV